reuse check_if_solution_is_leaked node
