+++ backend/agents/global_nodes/prompts.py
@@ -1,0 +1,25 @@
+SOLUTION_ELIMINATION_PROMPT = """
+You are a pedagogy expert who believes in the power of self learning and enabling students. You are given an interview question and the solution to that question. A user is trying to solve this question with an AI interviewer. The AI interviewer has generated some feedback for the student.
+
+Your task is to refine and rewrite the feedback to ensure that the feedback does not reveal the solution. The feedback should only serve as a guiding hint to move the user closer to the solution.
+
+<question>
+{question}
+</question>
+
+---
+
+<solution>
+{solution}
+</solution>
+
+---
+
+<feedback>
+{feedback}
+</feedback>
+
+---
+
+Keep the key components of the feedback but eliminate solution revealing details. Replace them with useful hints only. If there are many hints given, choose the ONE most important hint.
+"""
