+++ backend/agents/subgraphs/main_stage/prompts.py
@@ -75,34 +75,6 @@ Ground all your responses on the question and solution below.
 Remember to not reveal the solution.
 """
 
-SOLUTION_ELIMINATION_PROMPT = """
-You are a pedagogy expert who believes in the power of self learning and enabling students. You are given an interview question and the solution to that question. A user is trying to solve this question with an AI interviewer. The AI interviewer has generated some feedback for the student.
-
-Your task is to refine and rewrite the feedback to ensure that the feedback does not reveal the solution. The feedback should only serve as a guiding hint to move the user closer to the solution.
-
-<question>
-{question}
-</question>
-
----
-
-<solution>
-{solution}
-</solution>
-
----
-
-<feedback>
-{feedback}
-</feedback>
-
----
-
-Keep the key components of the feedback but eliminate solution revealing details. Replace them with useful hints only. If there are many hints given, choose the ONE most important hint.
-"""
-
-
-
 
 USER_INTENT_CLASSIFIER_PROMPT = """
 You are an expert at understanding user's intent. Detect the user's intent from the last message of the conversation.
